move rq helper to toolbox
